// src/app/openapi-client/model/productDetailDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CategoryShowDto } from './categoryShowDto';


export interface ProductDetailDto { 
    id: number;
    sku: string;
    active: boolean;
    name: string;
    image: string;
    description: string;
    price: number;
    stock: number;
    category?: CategoryShowDto;
}